uncertainty/requirements.txt
matplotlib==3.8.0
numpy==1.26.4
pytorch==2.2.2
scikit-learn==1.4.1
torchvision==0.17.2
tqdm==4.66.2